finished xaes usage; taproot test vector not passing due to tweak addition
