+++ frost-secp256k1/Cargo.toml
@@ -16,7 +16,6 @@ features = ["serde"]
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-xaes-256-gcm.workspace = true
 document-features.workspace = true
 frost-core.workspace = true
 frost-rerandomized.workspace = true
